requirements.txt
streamlit
google-generativeai
google-search-results
requests